+++ Model/Res.cs
@@ -1,14 +1,14 @@
 ﻿namespace ScootSim {
     struct Res {
-        public int Metal, Crystal, Deut;
+        public long Metal, Crystal, Deut;
 
-        public Res( int metal, int crystal, int deut ) {
+        public Res( long metal, long crystal, long deut ) {
             Metal = metal;
             Crystal = crystal;
             Deut = deut;
         }
 
-        public Res GetMultiple( int multiplier ) {
+        public Res GetMultiple( long multiplier ) {
             return new Res( Metal*multiplier/100, Crystal*multiplier/100, Deut*multiplier/100 );
         }
     }
